requirements.txt
jsonschema>=4.0.0
pydantic>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
openai>=1.0.0
anthropic>=0.7.0
pytest>=7.0.0
pytest-mock>=3.0.0
